+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- show project version (APP_VERSION env, else package.json) in status page footer. @vhidvz
+
 ## [1.0.0] - 2026-06-27
 
 ### Added